requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
